move HttpParameter class
