+++ SystemSprawozdan/SystemSprawozdan.Frontend/Services/AppHttpClient.cs
@@ -1,37 +1,9 @@
 ﻿using System.Text.Json;
 using System.Net.Http.Json;
+using SystemSprawozdan.Frontend.CustomClasses;
 
 namespace SystemSprawozdan.Frontend.Services
 {
-    public class HttpParameter {
-        public HttpParameter(string key, string value)
-        {
-            Key = key;
-            Value = value;
-        }
-        public string Key { get; set; }
-        public string Value { get; set; }
-
-        public string AddToUrl(string url)
-        {
-            if (!url.Contains("?"))
-                return $"{url}?{Key}={Value}";
-            
-            if(url.Contains($"{Key}="))
-            {
-                var start = url.IndexOf($"{Key}=");
-                var end = url.Substring(start).IndexOf("&");
-                end = end == -1 ? url.Length - start : end;
-                
-                var replaceFragment = url.Substring(start, end);
-                url = url.Replace(replaceFragment, $"{Key}={Value}");
-                return url;
-            }
-            
-            return $"{url}&{Key}={Value}";
-        }
-    }
-
     public interface IAppHttpClient
     {
         public Task<TResponse?> Get<TResponse>(string url, List<HttpParameter>? parameters = null);
