+++ SystemSprawozdan/SystemSprawozdan.Frontend/_Imports.razor
@@ -1,5 +1,6 @@
 ﻿@using System.Net.Http
 @using System.Net.Http.Json
+@using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
@@ -10,6 +11,7 @@
 @using MatBlazor
 
 @using SystemSprawozdan.Frontend
+@using SystemSprawozdan.Frontend.CustomClasses
 @using SystemSprawozdan.Frontend.Providers
 @using SystemSprawozdan.Frontend.Services
 @using SystemSprawozdan.Frontend.Shared
@@ -20,5 +22,4 @@
 @using SystemSprawozdan.Shared.Dto
 @using SystemSprawozdan.Shared.Enums
 
-@using Microsoft.AspNetCore.Authorization
 @attribute [Authorize]
